Big edits to xml.c, small edits to auto-host.h.in
  * GCC/config_cmake/auto-host.h.in
    - Fix GCC_INCLUDE_PATH in auto-host.h.in. This was set to
"bogus-include-dir" so that gccxml would use it's own Support includes.
Removing that definition from auto-host.h.in allowed the command-line
definitions to be accepted, which are now configured to specify the
system-wide GCC header directories.

  * xml.c
    - Incorporate Andrej Mitrovic's edits. These provide further support
for POD attributes and anonymous and type-scoped types. Change array.h
include to vec.h. Change a macro or two to their GCC 4.7 counterparts.